+++ app_modules/PedidosFarmacias/models/PedidosFarmaciasModel.js
@@ -27,6 +27,8 @@ PedidosFarmaciasModel.prototype.listar_farmacias_usuario = function(tipo, usuari
     }
     
     //Depreciado para el modulo de kardex, se traen las empresas desde el modulo de Empresas 24/06/2015
+    //Depreciado para el modulo de asignacion, se traen del perfil del usuario   25/06/2015
+    //Depreciado para el modulo de auditoria, se traen del perfil del usuario 25/06/2015
     if (tipo === '1') {
 
         sql = " SELECT\
@@ -879,7 +881,7 @@ PedidosFarmaciasModel.prototype.calcular_cantidad_total_pendiente_producto = fun
 
 PedidosFarmaciasModel.prototype.calcular_cantidad_reservada_temporales_farmacias = function(codigo_producto, callback) {
     
-    var sql = " select codigo_producto, SUM(cantidad_solic) as total_reservado from solicitud_pro_a_bod_prpal_tmp where codigo_producto = $1\
+    var sql = " select codigo_producto, SUM(cantidad_solic)::integer as total_reservado from solicitud_pro_a_bod_prpal_tmp where codigo_producto = $1\
                 group by codigo_producto"; 
     
     G.db.query(sql, [codigo_producto], function(err, rows, result) {
@@ -895,7 +897,7 @@ PedidosFarmaciasModel.prototype.calcular_cantidad_reservada_temporales_farmacias
 
 PedidosFarmaciasModel.prototype.calcular_cantidad_reservada_temporales_farmacias_por_fecha = function(codigo_producto, fecha_registro_pedido, callback) {
     
-    var sql = " select codigo_producto, SUM(cantidad_solic) as total_reservado from solicitud_pro_a_bod_prpal_tmp where codigo_producto = $1\
+    var sql = " select codigo_producto, SUM(cantidad_solic)::integer as total_reservado from solicitud_pro_a_bod_prpal_tmp where codigo_producto = $1\
                 and fecha_registro < $2\
                 group by codigo_producto ";
